vision/requirements.txt
torch
torchvision
transformers>=4.40.0
supervision>=0.19.0
fastapi
uvicorn[standard]
websockets
Pillow
numpy
opencv-python-headless
